Increase instrumentation of the match failures during testing.
Add cost information to the discrete MappingOp objects so we can track
the cost of an individual mapping step.

Fix String Jaccard distance computation s.t. we ignore zero-length
strings.  This improves match quality substantially.  On the Freebase
test regime, we now obtain MRR of >0.8 over ~1600 tests with the
attribute labels obscured.  The match operation now fails 17% of the
time, instead of 33%.  Of those failures, 80% are for 2-element
schemas, which we believe will be rare in real deployments.  This is
good enough that I think it is OK to move beyond the Freebase test set
to a more HDFS-centric one.